+++ src/eth/EthBalances.js
@@ -1,12 +1,12 @@
 'use strict';
 
 import React from 'react';
-import { Text } from 'ink';
+import {Text} from 'ink';
 import Table from 'ink-table';
-import { t } from 'saihubot-cli-adapter/dist/i18n';
+import {t} from 'saihubot-cli-adapter/dist/i18n';
 
-import useEthscanBalance from './useEthscanBalance';
-import useEthscanTokensBalance from './useEthscanTokensBalance';
+import useEthscanBalance from '../helpers/useEthscanBalance';
+import useEthscanTokensBalance from '../helpers/useEthscanTokensBalance';
 import {formatData} from '../helpers/format';
 
 import {i18nValidator, i18nBalance} from '../i18n';
@@ -17,17 +17,17 @@ export const EthBalances = ({addresses, fetch}) => {
   const [loadToken, balanceToken] = useEthscanTokensBalance(fetch, addresses);
 
   if (loadEth && loadToken) {
-    return (<Text>{t('query', {i18n: i18nBalance})}</Text>)
+    return (<Text>{t('query', {i18n: i18nBalance})}</Text>);
   }
 
   const balance = [...formatData(balanceEth), ...formatData(balanceToken)];
-  return balance.length > 0
-  ? (<>
+  return balance.length > 0 ?
+  (<>
     <Text>{t('accountBalance', {i18n: i18nValidator})}</Text>
     <Table data={balance} />
     <Text> </Text>
-  </>)
-  : null;
-}
+  </>) :
+  null;
+};
 
 export default EthBalances;